// src/app/bna-trainee-biodata/biodata-tab-layout/models/EducationalQualification.ts
export interface EducationalQualification {
    educationalQualificationId: number;
    traineeId: number;  
    examTypeId: number;  
    examType: string;  
    board: string;  
    group: string;  
    boardId: number;    
    address: string;
    subject: string;
    instituteName: string;
    groupId: number;
    passingYear: string;
    result: string;
    outOfResult: string;
    courseDuration: string;
    additionaInformation: string;
    status: number;
    menuPosition: number;
    isActive: boolean;
} 